feat: router with healthcheck.
