+++ package.json
@@ -5,7 +5,7 @@
     "main": "index.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "node --watch src/infra/db/sequelize/server.js"
+        "start": "node --watch src/app.js"
     },
     "keywords": [],
     "author": "",
